+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/EnchantmentsFeature.java
@@ -215,6 +215,7 @@ public class EnchantmentsFeature extends JsonFeature {
 			return;
 		if (attacker instanceof Player player)
 			knockback *= player.getAttackStrengthScale(0.5f);
+		knockback *= BonusDamageEnchantment.getDamageBonusRatio(attacker.getMainHandItem()) / 5f;
 		event.setStrength(event.getStrength() + knockback);
 	}