+++ changelog.md
@@ -4,6 +4,7 @@
 * If hit, Regenerating Absorption now goes on cooldown for 4 seconds instead of 2
 * Players now lose 20% xp on death instead of 15%
 * Bows and crossbows now have halved durability
+* Leaves are now twice as dull
 
 ## 3.2.1
 * Added `iguanatweaks:deathGrave` game rule